# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GatherUp API - Community events platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EventJoinMode = typeof EventJoinMode[keyof typeof EventJoinMode];


export const EventJoinMode = {
  open: 'open',
  approval_required: 'approval_required',
} as const;

export interface Event {
  id: number;
  title: string;
  description: string;
  category: string;
  location: string;
  date: string;
  time: string;
  hostId: number;
  /** @nullable */
  hostName?: string | null;
  /** @nullable */
  hostAvatar?: string | null;
  attendeeCount: number;
  maxAttendees: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  distanceKm?: number | null;
  joinMode?: EventJoinMode;
  createdAt: string;
}

export type EventInputJoinMode = typeof EventInputJoinMode[keyof typeof EventInputJoinMode];


export const EventInputJoinMode = {
  open: 'open',
  approval_required: 'approval_required',
} as const;

export interface EventInput {
  title: string;
  description: string;
  category: string;
  location: string;
  date: string;
  time: string;
  hostId: number;
  maxAttendees: number;
  imageUrl?: string;
  latitude?: number;
  longitude?: number;
  joinMode?: EventInputJoinMode;
}

export interface EventUpdate {
  title?: string;
  description?: string;
  category?: string;
  location?: string;
  date?: string;
  time?: string;
  maxAttendees?: number;
  imageUrl?: string;
}

export type EventStatsCategoryCountsItem = {
  category: string;
  count: number;
};

export interface EventStats {
  totalEvents: number;
  totalAttendees: number;
  categoryCounts: EventStatsCategoryCountsItem[];
}

export type AttendanceStatus = typeof AttendanceStatus[keyof typeof AttendanceStatus];


export const AttendanceStatus = {
  confirmed: 'confirmed',
  pending: 'pending',
  rejected: 'rejected',
} as const;

export interface Attendance {
  id: number;
  eventId: number;
  userId: number;
  status: AttendanceStatus;
  joinedAt: string;
}

export type JoinRequestStatus = typeof JoinRequestStatus[keyof typeof JoinRequestStatus];


export const JoinRequestStatus = {
  confirmed: 'confirmed',
  pending: 'pending',
  rejected: 'rejected',
} as const;

export interface JoinRequest {
  id: number;
  eventId: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userAvatar?: string | null;
  /** @nullable */
  userPhone?: string | null;
  status: JoinRequestStatus;
  joinedAt: string;
}

export interface AttendanceInput {
  userId: number;
}

export interface User {
  id: number;
  name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  profession?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  interests?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  instagram?: string | null;
  verified: boolean;
  /** @nullable */
  verifiedAt?: string | null;
  /** @nullable */
  verificationSelfieUrl?: string | null;
  /** @nullable */
  eventsHosted?: number | null;
  /** @nullable */
  eventsJoined?: number | null;
  createdAt: string;
}

export interface UserInput {
  name: string;
  bio?: string;
  avatarUrl?: string;
}

export interface UserUpdate {
  name?: string;
  bio?: string;
  avatarUrl?: string;
  profession?: string;
  age?: number;
  gender?: string;
  city?: string;
  interests?: string;
  email?: string;
  phone?: string;
  website?: string;
  instagram?: string;
}

export interface VerifyUserInput {
  selfieObjectPath: string;
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface Category {
  name: string;
  /** @nullable */
  icon?: string | null;
  eventCount: number;
}

export interface Comment {
  id: number;
  eventId: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userAvatar?: string | null;
  body: string;
  createdAt: string;
}

export interface CommentInput {
  userId: number;
  body: string;
}

export interface LoginInput {
  /** Phone number or email address */
  identifier: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthResponse {
  success: boolean;
  user?: User;
}

export interface RegisterInput {
  name: string;
  phone: string;
  email?: string;
  gender?: string;
  age?: number;
  profession?: string;
  /**
     * Account password (min 6 characters)
     * @minLength 6
     */
  password: string;
}

export interface Notification {
  id: number;
  type: string;
  title: string;
  /** @nullable */
  body?: string | null;
  read: boolean;
  createdAt: string;
  /** @nullable */
  actorId?: number | null;
  /** @nullable */
  actorName?: string | null;
  /** @nullable */
  actorAvatar?: string | null;
  /** @nullable */
  eventId?: number | null;
  /** @nullable */
  eventTitle?: string | null;
}

export interface UnreadCount {
  count: number;
}

export interface VapidPublicKey {
  publicKey: string;
}

export type PushSubscriptionInputKeys = {
  p256dh: string;
  auth: string;
};

export interface PushSubscriptionInput {
  endpoint: string;
  keys: PushSubscriptionInputKeys;
}

export interface PushUnsubscribeInput {
  endpoint: string;
}

export type ListEventsParams = {
category?: string;
search?: string;
upcoming?: boolean;
nearLat?: number;
nearLng?: number;
radiusKm?: number;
};