# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GatherUp API - Community events platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: upcoming
          in: query
          schema:
            type: boolean
        - name: nearLat
          in: query
          schema:
            type: number
        - name: nearLng
          in: query
          schema:
            type: number
        - name: radiusKm
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{id}/join:
    post:
      operationId: joinEvent
      tags: [events]
      summary: Join an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /events/{id}/leave:
    delete:
      operationId: leaveEvent
      tags: [events]
      summary: Leave an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "204":
          description: Left event
  /events/{id}/attendees:
    get:
      operationId: getEventAttendees
      tags: [events]
      summary: Get attendees for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of attendees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /events/{id}/requests:
    get:
      operationId: getEventJoinRequests
      tags: [events]
      summary: Get pending join requests for an event (organizer only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of join requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JoinRequest"
  /events/{id}/requests/{userId}/approve:
    post:
      operationId: approveJoinRequest
      tags: [events]
      summary: Approve a join request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /events/{id}/requests/{userId}/reject:
    post:
      operationId: rejectJoinRequest
      tags: [events]
      summary: Reject a join request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /events/suggested:
    get:
      operationId: getSuggestedEvents
      tags: [events]
      summary: Personalized event suggestions for the logged-in user
      responses:
        "200":
          description: Suggested events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    Event:
      type: object
      required: [id, title, description, category, location, date, time, hostId, attendeeCount, maxAttendees, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        location:
          type: string
        date:
          type: string
        time:
          type: string
        hostId:
          type: integer
        hostName:
          type: ["string", "null"]
        hostAvatar:
          type: ["string", "null"]
        attendeeCount:
          type: integer
        maxAttendees:
          type: integer
        imageUrl:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        distanceKm:
          type: ["number", "null"]
        joinMode:
          type: string
          enum: [open, approval_required]
        createdAt:
          type: string
    EventInput:
      type: object
      required: [title, description, category, location, date, time, hostId, maxAttendees]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        location:
          type: string
        date:
          type: string
        time:
          type: string
        hostId:
          type: integer
        maxAttendees:
          type: integer
        imageUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        joinMode:
          type: string
          enum: [open, approval_required]
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        location:
          type: string
        date:
          type: string
        time:
          type: string
        maxAttendees:
          type: integer
        imageUrl:
          type: string
...
    Attendance:
      type: object
      required: [id, eventId, userId, status, joinedAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [confirmed, pending, rejected]
        joinedAt:
          type: string
    JoinRequest:
      type: object
      required: [id, eventId, userId, status, joinedAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        userAvatar:
          type: ["string", "null"]
        userPhone:
          type: ["string", "null"]
        status:
          type: string
          enum: [confirmed, pending, rejected]
        joinedAt:
          type: string
    AttendanceInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
    User:
      type: object
      required: [id, name, verified, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        profession:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        city:
          type: ["string", "null"]
        interests:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        verified:
          type: boolean
        verifiedAt:
          type: ["string", "null"]
        verificationSelfieUrl:
          type: ["string", "null"]
        eventsHosted:
          type: ["integer", "null"]
        eventsJoined:
          type: ["integer", "null"]
        createdAt:
          type: string
...